benzinga_ws/requirements.txt
websockets>=12.0
google-cloud-logging>=3.10.0
